+++ plugins/dexie/src/tests/contract.test.ts
@@ -1,4 +1,4 @@
-import { describePluginContract, RENAMED_CALL_SELECTOR_TESTS } from '@routier/test-utils';
+import { describePluginContract } from '@routier/test-utils';
 import { uuidv4 } from '@routier/core';
 import { DexiePlugin } from '../DexiePlugin';
 
@@ -7,7 +7,6 @@ describePluginContract(
     () => new DexiePlugin(`contract-${uuidv4()}-db`),
     {
         supportsRichTypes: true,
-        knownFailing: RENAMED_CALL_SELECTOR_TESTS,
         reopen: plugin => new DexiePlugin(plugin.databaseName!),
     },
 );